test: move test files into tests directory and update imports
Separate test files and utilities from the `src` directory into `tests` for better organization.
