+++ lab/tests/integration/extras.test.ts
@@ -1,5 +1,5 @@
 import { test, expect } from "vitest";
-import { fetchContent } from "@/utils";
+import { fetchContent } from "../utils";
 
 test("hardbreak", async () => {
   const $ = await fetchContent("hardbreak");
